+++ addon/data-export.js
@@ -68,6 +68,7 @@ class Model {
   constructor({sfHost, args}) {
     this.sfHost = sfHost;
     this.queryInput = null;
+    this.filterColumn = ""; // Default filter column
     this.initialQuery = "";
     this.describeInfo = new DescribeInfo(this.spinFor.bind(this), () => {
       this.queryAutocompleteHandler({newDescribe: true});
@@ -1045,7 +1046,36 @@ function RecordTable(vm) {
       return "" + cell;
     }
   }
-  let isVisible = (row, filter) => !filter || row.some(cell => cellToString(cell).toLowerCase().includes(filter.toLowerCase()));
+
+  let isVisible = (row, filter) => {
+    // If no filter is applied, show all rows
+    if (!filter) {
+      return true;
+    }
+    // If no columns are selected, search all columns
+    if (!vm.filterColumns || vm.filterColumns.length === 0) {
+      return row.some(cell => {
+        if (cell == null) {
+          return false;
+        }
+        return cellToString(cell).toLowerCase().includes(filter.toLowerCase());
+      });
+    }
+
+    // Search in all selected columns
+    return vm.filterColumns.some(column => {
+      const columnIndex = header.findIndex(col => col === column);
+      if (columnIndex === -1) {
+        return false;
+      }
+
+      const cellValue = row[columnIndex];
+      return cellValue
+        ? cellToString(cellValue).toLowerCase().includes(filter.toLowerCase())
+        : false;
+    });
+  };
+
   let rt = {
     records: [],
     table: [],
@@ -1137,7 +1167,8 @@ class App extends React.Component {
     this.onResultsFilterInput = this.onResultsFilterInput.bind(this);
     this.onSetQueryName = this.onSetQueryName.bind(this);
     this.onStopExport = this.onStopExport.bind(this);
-    this.state = {hideButtonsOption: JSON.parse(localStorage.getItem("hideExportButtonsOption"))};
+    this.state = {hideButtonsOption: JSON.parse(localStorage.getItem("hideExportButtonsOption")), isDropdownOpen: false};// Tracks whether the dropdown is open
+    this.filterColumns = []; // Initialize as an empty array
   }
   onQueryAllChange(e) {
     let {model} = this.props;
@@ -1274,6 +1305,9 @@ class App extends React.Component {
   onResultsFilterInput(e) {
     let {model} = this.props;
     model.setResultsFilter(e.target.value);
+    if (e.target.value.length == 0){
+      this.setState({isDropdownOpen: false});
+    }
     model.didUpdate();
   }
   onSetQueryName(e) {
@@ -1446,9 +1480,9 @@ class App extends React.Component {
           h("div", {className: "autocomplete-header"},
             h("span", {}, model.autocompleteResults.title),
             h("div", {className: "flex-right"},
-              h("button", {tabIndex: 1, disabled: model.isWorking, onClick: this.onExport, title: "Ctrl+Enter / F5", className: "highlighted"}, "Run Export"),
-              this.displayButton("export-query") ? h("button", {tabIndex: 2, onClick: this.onCopyQuery, title: "Copy query url", className: "copy-id"}, "Export Query") : null,
-              h("button", {tabIndex: 3, onClick: this.onQueryPlan, title: "Run Query Plan"}, "Query Plan"),
+              h("button", {tabIndex: 1, disabled: model.isWorking, onClick: this.onExport, title: "Ctrl+Enter / F5", className: "highlighted button-margin"}, "Run Export"),
+              this.displayButton("export-query") ? h("button", {tabIndex: 2, onClick: this.onCopyQuery, title: "Copy query url", className: "copy-id button-margin"}, "Export Query") : null,
+              h("button", {tabIndex: 3, onClick: this.onQueryPlan, title: "Run Query Plan", className: "button-margin"}, "Query Plan"),
               h("a", {tabIndex: 4, className: "button", hidden: !model.autocompleteResults.sobjectName, href: model.showDescribeUrl(), target: "_blank", title: "Show field info for the " + model.autocompleteResults.sobjectName + " object"}, model.autocompleteResults.sobjectName + " Field Info"),
               h("button", {tabIndex: 5, href: "#", className: model.expandAutocomplete ? "toggle contract" : "toggle expand", onClick: this.onToggleExpand, title: "Show all suggestions or only the first line"},
                 h("div", {className: "button-icon"}),
@@ -1495,7 +1529,45 @@ class App extends React.Component {
             this.displayButton("delete")
               ? h("button", {disabled: !model.canDelete(), onClick: this.onDeleteRecords, title: "Open the 'Data Import' page with preloaded records to delete (< 20k records). 'Id' field needs to be queried", className: "delete-btn"}, "Delete Records") : null,
           ),
-          h("input", {placeholder: "Filter Results", type: "search", value: model.resultsFilter, onInput: this.onResultsFilterInput}),
+          h("div", {className: "filter-controls"},
+            h("div", {className: "unified-search-input"},
+              h("input", {
+                className: "filter-input",
+                placeholder: model.filterColumns?.length > 0
+                  ? `Filter by (${model.filterColumns.length})`
+                  : "Filter",
+                type: "search",
+                value: model.resultsFilter,
+                onInput: this.onResultsFilterInput
+              }),
+              h("button", {className: "toggle no-left-radius no-left-border" + (this.state.isDropdownOpen ? " contract" : " expand"), title: "Show More Filters", disabled: !model.exportedData, onClick: () => this.setState({isDropdownOpen: !this.state.isDropdownOpen})}, h("div", {className: "button-toggle-icon"})),
+              this.state.isDropdownOpen && h("div", {className: "dropdown-menu"},
+                model.exportedData?.table[0]
+                  ?.filter(column => column !== "_")
+                  .map(column =>
+                    h("div", {
+                      key: column,
+                      className: `dropdown-item ${model.filterColumns?.includes(column) ? "selected" : ""}`,
+                      onClick: () => {
+                        if (model.filterColumns?.includes(column)) {
+                          model.filterColumns = model.filterColumns.filter(c => c !== column);
+                        } else {
+                          model.filterColumns = [...(model.filterColumns || []), column];
+                        }
+                        model.setResultsFilter(model.resultsFilter);
+                        this.setState({}); // Trigger re-render
+                      }
+                    },
+                    h("input", {
+                      type: "checkbox",
+                      checked: model.filterColumns?.includes(column) || false,
+                      readOnly: true
+                    }),
+                    column
+                    )
+                  )
+              )
+            )),
           h("span", {className: "result-status flex-right"},
             h("span", {}, model.exportStatus),
             perf && h("span", {className: "result-info", title: perf.batchStats}, perf.text),